PART1/STEP3 grpc-api/nosql/requirements.txt
alembic==1.0.0
appdirs==1.4.3
astroid==2.0.1
attrs==18.2.0
bcrypt==3.1.4
black==18.6b4
bson==0.5.6
cffi==1.11.5
click==6.7
colorama==0.3.9
dominate==2.3.1
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.6
Flask-Login==0.4.1
Flask-Migrate==2.2.1
Flask-PyMongo==2.0.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
googleapis-common-protos==1.5.3
grpcio==1.15.0
grpcio-tools==1.15.0
gunicorn==19.9.0
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
protobuf==3.6.1
psycopg2==2.7.5
pycparser==2.18
pylint==2.0.1
pymongo==3.7.1
python-dateutil==2.7.3
python-editor==1.0.3
six==1.11.0
SQLAlchemy==1.2.10
toml==0.9.6
typed-ast==1.1.0
virtualenv==16.0.0
visitor==0.1.3
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.2.1
yapf==0.24.0
